requirements.txt
aiohttp
beautifulsoup4
certifi
charset-normalizer
colorama
idna
lxml
matplotlib
numpy
pandas
Pillow
python-dotenv
python-telegram-bot==13.15
requests
selenium
soupsieve
telethon
urllib3
undetected-chromedriver
